requirements.txt
streamlit
torch
transformers
accelerate
bitsandbytes
pypdf
python-docx
peft
